--- dd2dms.py
@@ -1,5 +1,0 @@
-def dd2dms(dd):
-    minutes, seconds = divmod(abs(dd) * 3600, 60)
-    degrees, minutes = divmod(minutes, 60)
-    dms = degrees + (minutes / 100) + (seconds / 10000)
-    return dms if dd >= 0 else -dms
