+++ llh2xyz.py
@@ -1,5 +1,26 @@
 from decimal import *
 from math import sqrt, sin, cos, radians
+from constants import grs80
+from conversions import dd2dms, dms2dd
+
+
+"""
+# Debug - Test Data
+lat = -31.515996736
+long = 133.483540489
+ellht = 144.7229
+lat = dms2dd(lat)
+long = dms2dd(long)
+"""
+
+
+# Universal Transverse Mercator Projection Parameters
+proj = grs80
+# Calculate Projection Constants
+f = float(1 / proj[1])
+a = proj[0]
+b = a * (1 - f)
+e2 = f * (2 - f)
 
 
 def llh2xyz(lat, long, ellht):
@@ -10,17 +31,9 @@ def llh2xyz(lat, long, ellht):
     # Convert lat & long to radians
     lat = radians(lat)
     long = radians(long)
-    # Universal Transverse Mercator Projection Parameters
-    Proj = [6378137, Decimal('298.25722210088'), 500000,
-            10000000, Decimal('0.9996'), 6, -177]
-    # Calculate Projection Constants
-    f = float(1/Proj[1])
-    a = Proj[0]
-    b = a*(1-f)
-    e2 = f*(2 - f)
     # Calculate Ellipsoid Radius of Curvature in the Prime Vertical - nu
     if lat == 0:
-        nu = Proj[0]
+        nu = proj[0]
     else:
         nu = a/(sqrt(1 - e2 * (sin(lat)**2)))
     # Calculate x, y, z
